Refactored for Contao 4
