+++ composer.json
@@ -20,24 +20,20 @@
     "source":"https://github.com/1up-lab/contao-j2b"
   },
   "require":{
-    "php":">=5.4",
-    "contao/core-bundle":"^3.2 || ^4.1",
-    "contao-community-alliance/composer-plugin":"~2.4 | ~3.0"
+    "php":">=7.0",
+    "contao/core-bundle":"^4.0",
+    "contao-community-alliance/composer-plugin":"^3.0"
+  },
+  "require-dev": {
+    "contao/manager-plugin": "^2.0"
   },
   "autoload": {
     "psr-4": {
-      "Oneup\\Contao\\": "src"
+      "Oneup\\Contao\\J2BBundle\\": "src"
     }
   },
   "extra": {
-    "contao": {
-      "sources": {
-        "contao": "system/modules/j2b"
-      }
-    },
-    "branch-alias": {
-      "dev-master": "1.0.x-dev"
-    }
+    "contao-manager-plugin": "Oneup\\Contao\\J2BBundle\\ContaoManager\\Plugin"
   },
   "minimum-stability": "dev",
   "prefer-stable": true
