--- contao/config/config.php
@@ -1,3 +1,0 @@
-<?php
-
-$GLOBALS['TL_HOOKS']['generatePage'][] = array('Oneup\Contao\J2B\Runner', 'moveJs');
