+++ src/ContaoManager/Plugin.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace Oneup\Contao\J2BBundle\ContaoManager;
+
+use Contao\CoreBundle\ContaoCoreBundle;
+use Contao\ManagerPlugin\Bundle\BundlePluginInterface;
+use Contao\ManagerPlugin\Bundle\Config\BundleConfig;
+use Contao\ManagerPlugin\Bundle\Parser\ParserInterface;
+use Oneup\Contao\J2BBundle\OneupContaoJ2BBundle;
+
+class Plugin implements BundlePluginInterface
+{
+    public function getBundles(ParserInterface $parser): array
+    {
+        return [
+            BundleConfig::create(OneupContaoJ2BBundle::class)->setLoadAfter([ContaoCoreBundle::class]),
+        ];
+    }
+
+}
