+++ src/DependencyInjection/OneupContaoJ2BExtension.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Oneup\Contao\J2BBundle\DependencyInjection;
+
+use Symfony\Component\Config\FileLocator;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Extension\Extension;
+use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
+
+class OneupContaoJ2BExtension extends Extension
+{
+    public function load(array $configs, ContainerBuilder $container): void
+    {
+        $loader = new YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
+        $loader->load('listener.yml');
+    }
+
+}
