+++ src/OneupContaoJ2BBundle.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Oneup\Contao\J2BBundle;
+
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+class OneupContaoJ2BBundle extends Bundle
+{
+}
