+++ src/Resources/config/listener.yml
@@ -1,0 +1,3 @@
+services:
+    oneup_contao_j2b.listener.js_rewrite:
+        class: Oneup\Contao\J2BBundle\EventListener\RewriteJsListener
