+++ src/Resources/contao/config/config.php
@@ -1,0 +1,3 @@
+<?php
+
+$GLOBALS['TL_HOOKS']['generatePage'][] = ['oneup_contao_j2b.listener.js_rewrite', 'onPageGenerate'];
